refactor(colocation): separate endpoint and result class
